devfreq_boost: Separate function to boost MIF when committing a new frame
This separates the MIF boosting done by 'decon' when the device shows a new frame, allowing us to consistently set a boost duration from Kconfig with the default being 70ms. This also allows us to make the feature configurable from sysfs, which we will modify in userspace to enable the feature at runtime.

This also allows us to allow others to boost MIF even if there is video playing in the background, i.e. when Zygote is forked or fingerprint authentication is in place.

Test for functionality and potential additional overhead.

Signed-off-by: John Vincent <git@tenseventyseven.cf>