sched: exynos: Remove preemption check for fit CPU selection
We can now remove and skip this process as Task EXpress (TEX) already excludes CPUs it assigned TEX tasks to, with Mint's implementation including the task(s) for the program currently shown to the user, which this preemption check formerly did. TEX' current implementation seen on Android 13 kernels also excludes fast CPUs for TEX-'suppressed' tasks, completely making this legacy check unneeded at this point.

This removes the overhead of checking the runqueue and whether the current task in the rq is boosted, and we can finally remove wake_cap() this time.

Signed-off-by: John Vincent <git@tenseventyseven.cf>