sched: exynos: Make 'mulligan' tasks ineligible for fast placement
Ontime's 'mulligan' migration aims to move small tasks on the current, busy CPU to a free one. While tasks migrated in this manner will very likely be ineligible for fast placement due to the migration path's nature, check for it anyway so the checks below would no longer be needed for these tasks.

Signed-off-by: John Vincent <git@tenseventyseven.cf>