sched: exynos: Allow semi-perf policy to all cgroups that get KPP hints
Remove the ambiguity of tasks in the 'top-app' cgroup being the only one being scheduled as 'semi-perf' when kernel "prefer-perf" hints are true for others, usually for tasks in the 'foreground' cgroup as well. Allow all tasks that get the kernel's "prefer-perf" hint to be scheduled as 'semi-perf' as well, not just 'top-app' tasks.

This shouldn't be an energy compromise for the most part as we have tweaked the 'semi-perf' path to consider the active CPU instead when the task is not latency-sensitive, which the 'foreground' cgroup is set not to by default. This may also improve performance during screen touches and app switches.

Signed-off-by: John Vincent <git@tenseventyseven.cf>