sched: exynos: Drop EMS 'frequency variant' boosting
This is part of an effort to drop legacy components of the Exynos Mobile Scheduler (EMS), preferring more modern solutions. Completely drop EMS' frequency variant boosting; this is a noticeable overhead that we honestly don't really need as we have SchedTune handling CPU boosting solely post-UtilClamp merger.

If we ever need a feature from this, just add it to AIGo.

Signed-off-by: John Vincent <git@tenseventyseven.cf>