devfreq_boost: Boost MIF based on EMS SysBusy indicators
On the upstream Exynos devfreq driver, MIF is boosted to the max when SysBusy indicators detect at least one busy CPU with heavy task(s) contributing much to the system's 'business'. Do the same for us in the platform, but take advantage of the "devfreq_boost" driver and hook it from here instead.

Signed-off-by: John Vincent <git@tenseventyseven.cf>